new commit

+++ frontend/src/components/SignupScreen.js
@@ -5,30 +5,38 @@ import { useContext, useEffect, useState } from 'react';
 import { Store } from '../Store';
 import { toast } from 'react-toastify';
 import { getError } from '../utils';
-export default function SigninScreen() {
+export default function SignupScreen() {
   const { search } = useLocation();
   const redirectInUrl = new URLSearchParams(search).get('redirect');
   const redirect = redirectInUrl ? redirectInUrl : '/';
 
+  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
   const { state, dispatch: ctxDispatch } = useContext(Store);
   const { userInfo } = state;
   const navigate = useNavigate();
   const submitHandler = async (e) => {
     e.preventDefault();
+    if (password !== confirmPassword) {
+      toast.error('Passwords do not match');
+      return;
+    }
     try {
-      const { data } = await Axios.post('/api/users/signin', {
+      const { data } = await Axios.post('/api/users/signup', {
+        name,
         email,
         password,
       });
       ctxDispatch({ type: 'USER_SIGNIN', payload: data });
       localStorage.setItem('userInfo', JSON.stringify(data));
       navigate(redirect || '/');
-      toast.success("Logged in successfully!");
+      toast.success("Account Created Successfully!");
     } catch (err) {
       toast.error(getError(err));
     }
+    
   };
 
   useEffect(() => {
@@ -40,12 +48,23 @@ export default function SigninScreen() {
   return (
     <div className="max-w-[600px]">
       <Helmet>
-        <title>Sign in</title>
+        <title>Sign Up</title>
       </Helmet>
-      <h1 className="my-3 text-3xl font-bold">Log in to your account</h1>
+      <h1 className="my-3 text-3xl font-bold">Create Your New Account</h1>
       <form onSubmit={submitHandler}>
         <div className="mb-3 flex flex-col">
-          <label className="text-gray-600">Email</label>
+          <label>Name</label>
+          <input
+            className="max-w-[400px] border-b-2 outline-none border-gray-200"
+            type="text"
+            name="name"
+            placeholder=""
+            required
+            onChange={(e) => setName(e.target.value)}
+          />
+        </div>
+        <div className="mb-3 flex flex-col">
+          <label>Email</label>
           <input
             className="max-w-[400px] border-b-2 outline-none border-gray-200"
             type="email"
@@ -56,7 +75,7 @@ export default function SigninScreen() {
           />
         </div>
         <div className="mb-3 flex flex-col">
-          <label className="text-gray-600">Password</label>
+          <label>Password</label>
           <input
             className="max-w-[400px] border-b-2 outline-none border-gray-200"
             type="password"
@@ -66,21 +85,32 @@ export default function SigninScreen() {
             onChange={(e) => setPassword(e.target.value)}
           />
         </div>
+        <div className="mb-3 flex flex-col">
+          <label>Confirm Password</label>
+          <input
+            className="max-w-[400px] border-b-2 outline-none border-gray-200"
+            type="password"
+            name="confirmpassword"
+            placeholder=""
+            required
+            onChange={(e) => setConfirmPassword(e.target.value)}
+          />
+        </div>
         <div className="mb-3 max-w-[400px] flex items-center justify-center ">
           <button
             className="bg-gray-800 p-1 w-[100px]  shadow-md font-semibold border-2 text-gray-200 hover:bg-gray-900"
             type="submit"
           >
-            Sign in
+            Sign Up
           </button>
         </div>
         <div className="mb-3 flex flex-col text-lg">
-          New Customer?{' '}
+          Already have an Account?{' '}
           <Link
             className="underline decoration-solid"
-            to={`/signup?redirect=${redirect}`}
+            to={`/signin?redirect=${redirect}`}
           >
-            Create your account
+            Sign-In
           </Link>
         </div>
       </form>
